docs: optimize README and architecture docs for open-source community
- Clarify service positioning as local daemon runtime with OpenAI compatibility
- Refine offline privacy-first descriptions and remove jargon
- Clarify float16 and int8 dtype configuration and memory optimization choices
- Add Qwen3-ASR (0.6B/1.7B) and Qwen3-TTS supported model matrices and guide
- Restructure architecture diagrams into concise overview and detailed topology

+++ src/speechrail/http/routes/audio.py
@@ -149,7 +149,7 @@ def _try_fast_decode_wav(audio: bytes) -> bytes | None:
                 samples = samples.reshape(-1, 2).mean(axis=1)
 
             if sample_rate != 16_000 and sample_rate is not None and sample_rate > 0:
-                num_out = int(round(len(samples) * 16_000.0 / sample_rate))
+                num_out = round(len(samples) * 16_000.0 / sample_rate)
                 if num_out <= 0:
                     return None
                 x_old = np.arange(len(samples), dtype=np.float32)
